Added CLI/Environment/Config File Parameters
The hardest part was figuring out that I could actually modify the os.Args
slice and add in a value if I wanted to. This allowed me to add in an
extra feature on top of the Flags package I added in that had the
capability to pull from the Environment Variables or a Configuration File.
I also wanted the capability for the program to detect if a configuration
file was present and use it automatically without needing to specify a
command line flag.

This allowed me to remove the GetPort function we initially had in here
that we had brought in from some example code we had started from.

Additionally, an HTTP MUX was added in so that the favicon images I
generated can be served from the root folder, since most web browsers
expect the favicon to be located in the root.

Some extra link elements were added into for all of the icon images into
our basic HTML page, but this will be expanded in one of the upcoming
commits as we move to using actual HTML templates.