Started integrating the HTML Template
Finally figured out what the issue was preventing the /assets URL
path from working correctly...the third parameter pointing to the
/public/assets URL directory needs to be written without the initial
forward slash because if it's included it doesn't look at the path as a
relative one. I should have noticed that sooner, but it took longer than
expected.

Created a templates folder and started creating what will be our HTML
template for the site.

In doing so I did introduce an issue though...currently our UserInfo
struct doesn't seem to passing its values correctly to the template so I
need to figure out why.

After we get that sorted out, then it'll mainly be starting to customize
the HTML template and getting it fully presentable.