Fixed issue with template parameters not showing
The IP Address / Hostname values weren't showing up on the sublayout.

It appears that you need to add a "." to reference the passed in data
object when calling another template from the main template so that it
will have access to the data too.

Started a small portion of the HTML cleanup by removing some of the
unneeded HTML.